loan report and user image show in side bar

+++ app/Http/Controllers/Auth/UserController.php
@@ -96,7 +96,7 @@ class UserController extends Controller
 
         if ($validatedData['image']) {
 
-            if ($validatedData['old_image']) {
+            if (isset($validatedData['old_image']) && $validatedData['old_image']) {
                 removeFile($validatedData['old_image']);
             }
